Helper functions with Total month quality
